+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 - [FSharp] Added NativePtr.stackUseArr
 - [Report] Using Stopwatch.GetTimestamp() instead of Stopwatch instance in ReportJob
+- [FSharp] Added Map.tryFindV
+- [FSharp] Added Map.ofSeqWithDuplicates
+- [FSharp] Added choosei, collecti for Seq, List, Array
 
 ### 5.3.8
 - [FSharp] Added stableSum and stableSumBy for List, Array, and Seq